Extract the label applicator
